\name{GeneSets}
\docType{data}
\alias{GeneSets}
\alias{MSIG.geneSets}
\alias{ISG.geneSet}
\alias{BTM.geneSets}
\title{Example Gene Sets}
\description{
  \describe{
  \item{ISG.geneSet}{A vector contains a list of probable ISGs based on Schoggins et al. (Nature, 2011).}
  \item{MSIG.geneSet}{A list containing a set of vectors representing the MSigDB's Canonical Pathways gene set database.}
  \item{BTM.geneSet}{A list containing a set of vectors. Each entry describes one of the blood transcription modules (BTMs) from Li et al (Nat Immunol., 2014)}
  }
}
\usage{
  ISG.geneSet
  MSIG.geneSets
  BTM.geneSet
}
\format{}
\source{\url{http://www.broadinstitute.org/gsea/msigdb/}}
\references{
 ISG: Schoggins et al. (Nature, 2011).
 
 MSIG: Liberzon et al. (Bioinformatics, 2011)
 
 BTM: Li et al. (Nat Immunol., 2014).
}
\keyword{datasets}