\name{ISG.geneSet}
\docType{data}
\alias{ISG.geneSet}
\title{Example gene set containing IFN stimulated genes}
\description{
 This vector contains a list probable ISGs based on Schoggins et al. (Nature, 2011) 
 and validated by ______.
}
\usage{ISG.geneSet}
\format{A vector containing 227 gene symbols.}
\source{various}
\references{
 Schoggins et al. (Nature, 2011).
}
\keyword{datasets}