\name{MSIG.geneSets}
\docType{data}
\alias{MSIG.geneSets}
\title{MSigDB's Canonical Pathways gene set database.}
\description{
 This is a list containing a set of vectors representing the MSigDB's Canonical Pathways gene set database.
}
\usage{MSIG.geneSets}
\format{A list containing 227 gene symbols.}
\source{\url{http://www.broadinstitute.org/gsea/msigdb/}}
\references{
 Liberzon et al. (Bioinformatics, 2011)
}
\keyword{datasets}